# Replace ReadStringAsSlice with ReadRawString