--- a/iter_skip_strict.go
+++ b/iter_skip_strict.go
@@ -1,4 +1,5 @@
-//+build !jsoniter_sloppy
+//go:build !jsoniter_sloppy
+// +build !jsoniter_sloppy
 
 package jsoniter
 
@@ -72,12 +73,12 @@ func (iter *Iterator) skipString() {
 		}
 	}
 
-	iter.readStringInner()
+	iter.readRawStringInner()
 }
 
 func (iter *Iterator) skipObject() {
 	iter.unreadByte()
-	iter.ReadObjectCB(func(iter *Iterator, field string) bool {
+	iter.ReadObjectCB(func(i *Iterator, s string) bool {
 		iter.Skip()
 		return true
 	})