This also allowed more efficient iteration of object keys, so there's now also ReadObjectRaw + CB

benchmarks against master:
```
name                                 old time/op    new time/op    delta
_encode_string_with_SetEscapeHTML-8     498ns ± 0%     498ns ± 0%     ~     (p=0.444 n=5+5)
_jsoniter_large_file-8                 9.88µs ± 2%    9.24µs ± 1%   -6.40%  (p=0.008 n=5+5)
_json_large_file-8                     23.9µs ± 0%    24.0µs ± 1%     ~     (p=1.000 n=5+5)
DatasetCanada/bytes-8                  15.8ms ± 0%    15.9ms ± 2%     ~     (p=0.730 n=4+5)
DatasetGithub/bytes-8                   116µs ± 0%     102µs ± 1%  -11.99%  (p=0.016 n=5+4)
DatasetGsoc2018/bytes-8                4.73ms ± 1%    4.11ms ± 1%  -12.94%  (p=0.008 n=5+5)
DatasetTwitterEscaped/bytes-8          1.95ms ± 1%    1.80ms ± 1%   -7.69%  (p=0.008 n=5+5)
ReadLongString/simple-8                9.69µs ± 1%    7.87µs ± 0%  -18.76%  (p=0.008 n=5+5)
ReadLongString/escaped-8               14.4µs ± 6%    12.5µs ± 2%  -13.26%  (p=0.008 n=5+5)
_stream_encode_big_object-8            4.92µs ± 1%    4.90µs ± 0%     ~     (p=0.690 n=5+5)

name                                 old alloc/op   new alloc/op   delta
_encode_string_with_SetEscapeHTML-8      760B ± 0%      760B ± 0%     ~     (all equal)
_jsoniter_large_file-8                 4.91kB ± 0%    4.37kB ± 0%  -11.07%  (p=0.008 n=5+5)
_json_large_file-8                     10.4kB ± 0%    10.4kB ± 0%     ~     (all equal)
DatasetCanada/bytes-8                  2.61MB ± 0%    2.61MB ± 0%   -0.00%  (p=0.008 n=5+5)
DatasetGithub/bytes-8                  65.2kB ± 0%    55.3kB ± 0%  -15.24%  (p=0.008 n=5+5)
DatasetGsoc2018/bytes-8                6.38MB ± 0%    6.23MB ± 0%   -2.47%  (p=0.008 n=5+5)
DatasetTwitterEscaped/bytes-8           587kB ± 0%     383kB ± 0%  -34.79%  (p=0.008 n=5+5)
ReadLongString/simple-8                14.3kB ± 0%    14.3kB ± 0%     ~     (all equal)
ReadLongString/escaped-8               43.4kB ± 9%    47.8kB ± 2%  +10.22%  (p=0.016 n=5+5)
_stream_encode_big_object-8             0.00B          0.00B          ~     (all equal)

name                                 old allocs/op  new allocs/op  delta
_encode_string_with_SetEscapeHTML-8      6.00 ± 0%      6.00 ± 0%     ~     (all equal)
_jsoniter_large_file-8                   78.0 ± 0%       5.0 ± 0%  -93.59%  (p=0.008 n=5+5)
_json_large_file-8                       16.0 ± 0%      16.0 ± 0%     ~     (all equal)
DatasetCanada/bytes-8                    109k ± 0%      109k ± 0%   -0.01%  (p=0.008 n=5+5)
DatasetGithub/bytes-8                   1.90k ± 0%     0.77k ± 0%  -59.82%  (p=0.008 n=5+5)
DatasetGsoc2018/bytes-8                 38.7k ± 0%     19.8k ± 0%  -48.95%  (p=0.008 n=5+5)
DatasetTwitterEscaped/bytes-8           20.6k ± 0%      7.4k ± 0%  -64.26%  (p=0.008 n=5+5)
ReadLongString/simple-8                  2.00 ± 0%      2.00 ± 0%     ~     (all equal)
ReadLongString/escaped-8                 12.0 ± 8%      13.4 ±12%     ~     (p=0.103 n=5+5)
_stream_encode_big_object-8              0.00           0.00          ~     (all equal)

name                                 old speed      new speed      delta
DatasetCanada/bytes-8                 135MB/s ± 0%   135MB/s ± 2%     ~     (p=0.730 n=4+5)
DatasetGithub/bytes-8                 537MB/s ± 0%   611MB/s ± 1%  +13.63%  (p=0.016 n=5+4)
DatasetGsoc2018/bytes-8               672MB/s ± 1%   772MB/s ± 1%  +14.86%  (p=0.008 n=5+5)
DatasetTwitterEscaped/bytes-8         275MB/s ± 1%   298MB/s ± 1%   +8.32%  (p=0.008 n=5+5)
```